requirements.txt
pydantic==2.9.2
python-decouple==3.8